style(cli): use appropriate icons per platform
